API Reference

NarrativeTest.runtests — Function

runtests(files; subs=common_subs(), mod=nothing, quiet=false) :: Bool

Loads the specified Markdown files to extract and run the embedded test cases. When a directory is passed, loads all *.md files in the directory. Returns true if the testing is successful, false otherwise.

Specify subs to customize substitutions applied to the expected output in order to convert it to a regular expression.

Specify mod to execute tests in the context of the given module.

Set quiet=true to suppress all output except for error reports.

runtests(; default=common_args(), subs=common_subs(), mod=nothing, quiet=false)

In this form, test files are specified as command-line parameters. When invoked without parameters, loads all *.md files in the program directory, which can be overriden using default parameter. Exits with code 0 if the testing is successful, 1 otherwise.

Use this form in test/runtests.jl:

using NarrativeTest
runtests()

NarrativeTest.runtest — Function

runtest(test::Test; subs=common_subs(), mod=nothing) :: AbstractResult
runtest(loc, code; pre=nothing, expect=nothing, subs=common_subs(), mod=nothing) :: AbstractResult

Runs the given test case, returns the result.

NarrativeTest.parsemd — Function

parsemd(file) :: Vector{AbstractTest}
parsemd(name, io) :: Vector{AbstractTest}

Parses the specified Markdown file to extract the embedded test suite. Returns a list of test cases.

NarrativeTest.parsejl — Function

parsejl(file) :: Vector{AbstractTest}
parsejl(name, io) :: Vector{AbstractTest}

Loads the specified Julia source file and extracts the embedded test suite. Returns a list of test cases.

NarrativeTest.common_args — Function

common_args() :: Vector{String}

Default test files for use when the test runner has no arguments.

NarrativeTest.common_subs — Function

common_subs() :: Vector{Pair{Regex,SubstitutionString{String}}}

Substitutions applied to the expected output in order to convert it to a regular expression.