API Reference

API Reference

NarrativeTest.runtests — Function.

runtests(files) :: Bool

Loads the specified Markdown files to extract and run the embedded test cases. When a directory is passed, loads all *.md files in the directory. Returns true if the testing is successful, false otherwise.

runtests()

In this form, test files are specified as command-line parameters. When invoked without parameters, loads all *.md files in the program directory. Exits with code 0 if the testing is successful, 1 otherwise. Use this form in test/runtests.jl:

using NarrativeTest
runtests()

source

NarrativeTest.runtest — Function.

runtest(test::Test) :: AbstractResult
runtest(loc, code, expect) :: AbstractResult

Runs the given test case, returns the result.

source

NarrativeTest.parsemd — Function.

parsemd(file) :: Vector{AbstractTest}
parsemd(name, io) :: Vector{AbstractTest}

Parses the specified Markdown file to extract the embedded test suite. Returns a list of test cases.

source

NarrativeTest.parsejl — Function.

parsejl(file) :: Vector{AbstractTest}
parsejl(name, io) :: Vector{AbstractTest}

Loads the specified Julia source file and extracts the embedded test suite. Returns a list of test cases.

source